+++ cluster/query_executor_test.go
@@ -2,6 +2,7 @@ package cluster_test
 
 import (
 	"bytes"
+	"errors"
 	"io"
 	"os"
 	"reflect"
@@ -70,6 +71,49 @@ func TestQueryExecutor_ExecuteQuery_SelectStatement(t *testing.T) {
 					{time.Unix(1, 0).UTC(), float64(200)},
 				},
 			}},
+		},
+	}) {
+		t.Fatalf("unexpected results: %s", spew.Sdump(a))
+	}
+}
+
+// Ensure query executor can enforce a maximum series selection count.
+func TestQueryExecutor_ExecuteQuery_MaxSelectSeriesN(t *testing.T) {
+	e := DefaultQueryExecutor()
+	e.MaxSelectSeriesN = 3
+
+	// The meta client should return a two shards on the local node.
+	e.MetaClient.ShardsByTimeRangeFn = func(sources influxql.Sources, tmin, tmax time.Time) (a []meta.ShardInfo, err error) {
+		return []meta.ShardInfo{
+			{ID: 100, Owners: []meta.ShardOwner{{NodeID: 0}}},
+			{ID: 101, Owners: []meta.ShardOwner{{NodeID: 0}}},
+		}, nil
+	}
+
+	// This iterator creator returns an iterator that operates on 2 series.
+	// Reuse this iterator for both shards. This brings the total series count to 4.
+	var ic IteratorCreator
+	ic.CreateIteratorFn = func(opt influxql.IteratorOptions) (influxql.Iterator, error) {
+		return &FloatIterator{
+			Points: []influxql.FloatPoint{{Name: "cpu", Time: int64(0 * time.Second), Aux: []interface{}{float64(100)}}},
+			stats:  influxql.IteratorStats{SeriesN: 2},
+		}, nil
+	}
+	ic.FieldDimensionsFn = func(sources influxql.Sources) (fields, dimensions map[string]struct{}, err error) {
+		return map[string]struct{}{"value": struct{}{}}, nil, nil
+	}
+	ic.SeriesKeysFn = func(opt influxql.IteratorOptions) (influxql.SeriesList, error) {
+		return influxql.SeriesList{
+			{Name: "cpu", Aux: []influxql.DataType{influxql.Float}},
+		}, nil
+	}
+	e.TSDBStore.ShardIteratorCreatorFn = func(id uint64) influxql.IteratorCreator { return &ic }
+
+	// Verify all results from the query.
+	if a := ReadAllResults(e.ExecuteQuery(`SELECT count(value) FROM cpu`, "db0", 0)); !reflect.DeepEqual(a, []*influxql.Result{
+		{
+			StatementID: 0,
+			Err:         errors.New("max select series count exceeded: 4 series"),
 		},
 	}) {
 		t.Fatalf("unexpected results: %s", spew.Sdump(a))
@@ -233,9 +277,10 @@ func (ic *IteratorCreator) ExpandSources(sources influxql.Sources) (influxql.Sou
 // FloatIterator is a represents an iterator that reads from a slice.
 type FloatIterator struct {
 	Points []influxql.FloatPoint
+	stats  influxql.IteratorStats
 }
 
-func (itr *FloatIterator) Stats() influxql.IteratorStats { return influxql.IteratorStats{} }
+func (itr *FloatIterator) Stats() influxql.IteratorStats { return itr.stats }
 func (itr *FloatIterator) Close() error                  { return nil }
 
 // Next returns the next value and shifts it off the beginning of the points slice.