+++ tsdb/engine/tsm1/iterator.gen.go
@@ -9,6 +9,7 @@ package tsm1
 import (
 	"fmt"
 	"sort"
+	"sync"
 
 	"github.com/influxdata/influxdb/influxql"
 	"github.com/influxdata/influxdb/tsdb"
@@ -97,6 +98,11 @@ func (c *bufCursor) nextAt(seek int64) interface{} {
 	}
 }
 
+// statsBufferCopyIntervalN is the number of points that are read before
+// copying the stats buffer to the iterator's stats field. This is used to
+// amortize the cost of using a mutex when updating stats.
+const statsBufferCopyIntervalN = 100
+
 type floatIterator struct {
 	cur   floatCursor
 	aux   []cursorAt
@@ -109,7 +115,9 @@ type floatIterator struct {
 	m     map[string]interface{} // map used for condition evaluation
 	point influxql.FloatPoint    // reusable buffer
 
-	stats influxql.IteratorStats
+	statsLock sync.Mutex
+	stats     influxql.IteratorStats
+	statsBuf  influxql.IteratorStats
 }
 
 func newFloatIterator(name string, tags influxql.Tags, opt influxql.IteratorOptions, cur floatCursor, aux []cursorAt, conds []*bufCursor, condNames []string) *floatIterator {
@@ -121,10 +129,11 @@ func newFloatIterator(name string, tags influxql.Tags, opt influxql.IteratorOpti
 			Name: name,
 			Tags: tags,
 		},
-		stats: influxql.IteratorStats{
+		statsBuf: influxql.IteratorStats{
 			SeriesN: 1,
 		},
 	}
+	itr.stats = itr.statsBuf
 
 	if len(aux) > 0 {
 		itr.point.Aux = make([]interface{}, len(aux))
@@ -160,10 +169,13 @@ func (itr *floatIterator) Next() *influxql.FloatPoint {
 
 		// Exit if we have no more points or we are outside our time range.
 		if itr.point.Time == tsdb.EOF {
+			itr.copyStats()
 			return nil
 		} else if itr.opt.Ascending && itr.point.Time > itr.opt.EndTime {
+			itr.copyStats()
 			return nil
 		} else if !itr.opt.Ascending && itr.point.Time < itr.opt.StartTime {
+			itr.copyStats()
 			return nil
 		}
 
@@ -183,14 +195,28 @@ func (itr *floatIterator) Next() *influxql.FloatPoint {
 		}
 
 		// Track points returned.
-		itr.stats.PointN++
+		itr.statsBuf.PointN++
+
+		// Copy buffer to stats periodically.
+		if itr.statsBuf.PointN%statsBufferCopyIntervalN == 0 {
+			itr.copyStats()
+		}
 
 		return &itr.point
 	}
 }
 
+// copyStats copies from the itr stats buffer to the stats under lock.
+func (itr *floatIterator) copyStats() {
+}
+
 // Stats returns stats on the points processed.
-func (itr *floatIterator) Stats() influxql.IteratorStats { return itr.stats }
+func (itr *floatIterator) Stats() influxql.IteratorStats {
+	itr.statsLock.Lock()
+	stats := itr.stats
+	itr.statsLock.Unlock()
+	return stats
+}
 
 // Close closes the iterator.
 func (itr *floatIterator) Close() error { return nil }
@@ -452,7 +478,9 @@ type integerIterator struct {
 	m     map[string]interface{} // map used for condition evaluation
 	point influxql.IntegerPoint  // reusable buffer
 
-	stats influxql.IteratorStats
+	statsLock sync.Mutex
+	stats     influxql.IteratorStats
+	statsBuf  influxql.IteratorStats
 }
 
 func newIntegerIterator(name string, tags influxql.Tags, opt influxql.IteratorOptions, cur integerCursor, aux []cursorAt, conds []*bufCursor, condNames []string) *integerIterator {
@@ -464,10 +492,11 @@ func newIntegerIterator(name string, tags influxql.Tags, opt influxql.IteratorOp
 			Name: name,
 			Tags: tags,
 		},
-		stats: influxql.IteratorStats{
+		statsBuf: influxql.IteratorStats{
 			SeriesN: 1,
 		},
 	}
+	itr.stats = itr.statsBuf
 
 	if len(aux) > 0 {
 		itr.point.Aux = make([]interface{}, len(aux))
@@ -503,10 +532,13 @@ func (itr *integerIterator) Next() *influxql.IntegerPoint {
 
 		// Exit if we have no more points or we are outside our time range.
 		if itr.point.Time == tsdb.EOF {
+			itr.copyStats()
 			return nil
 		} else if itr.opt.Ascending && itr.point.Time > itr.opt.EndTime {
+			itr.copyStats()
 			return nil
 		} else if !itr.opt.Ascending && itr.point.Time < itr.opt.StartTime {
+			itr.copyStats()
 			return nil
 		}
 
@@ -526,14 +558,28 @@ func (itr *integerIterator) Next() *influxql.IntegerPoint {
 		}
 
 		// Track points returned.
-		itr.stats.PointN++
+		itr.statsBuf.PointN++
+
+		// Copy buffer to stats periodically.
+		if itr.statsBuf.PointN%statsBufferCopyIntervalN == 0 {
+			itr.copyStats()
+		}
 
 		return &itr.point
 	}
 }
 
+// copyStats copies from the itr stats buffer to the stats under lock.
+func (itr *integerIterator) copyStats() {
+}
+
 // Stats returns stats on the points processed.
-func (itr *integerIterator) Stats() influxql.IteratorStats { return itr.stats }
+func (itr *integerIterator) Stats() influxql.IteratorStats {
+	itr.statsLock.Lock()
+	stats := itr.stats
+	itr.statsLock.Unlock()
+	return stats
+}
 
 // Close closes the iterator.
 func (itr *integerIterator) Close() error { return nil }
@@ -795,7 +841,9 @@ type stringIterator struct {
 	m     map[string]interface{} // map used for condition evaluation
 	point influxql.StringPoint   // reusable buffer
 
-	stats influxql.IteratorStats
+	statsLock sync.Mutex
+	stats     influxql.IteratorStats
+	statsBuf  influxql.IteratorStats
 }
 
 func newStringIterator(name string, tags influxql.Tags, opt influxql.IteratorOptions, cur stringCursor, aux []cursorAt, conds []*bufCursor, condNames []string) *stringIterator {
@@ -807,10 +855,11 @@ func newStringIterator(name string, tags influxql.Tags, opt influxql.IteratorOpt
 			Name: name,
 			Tags: tags,
 		},
-		stats: influxql.IteratorStats{
+		statsBuf: influxql.IteratorStats{
 			SeriesN: 1,
 		},
 	}
+	itr.stats = itr.statsBuf
 
 	if len(aux) > 0 {
 		itr.point.Aux = make([]interface{}, len(aux))
@@ -846,10 +895,13 @@ func (itr *stringIterator) Next() *influxql.StringPoint {
 
 		// Exit if we have no more points or we are outside our time range.
 		if itr.point.Time == tsdb.EOF {
+			itr.copyStats()
 			return nil
 		} else if itr.opt.Ascending && itr.point.Time > itr.opt.EndTime {
+			itr.copyStats()
 			return nil
 		} else if !itr.opt.Ascending && itr.point.Time < itr.opt.StartTime {
+			itr.copyStats()
 			return nil
 		}
 
@@ -869,14 +921,28 @@ func (itr *stringIterator) Next() *influxql.StringPoint {
 		}
 
 		// Track points returned.
-		itr.stats.PointN++
+		itr.statsBuf.PointN++
+
+		// Copy buffer to stats periodically.
+		if itr.statsBuf.PointN%statsBufferCopyIntervalN == 0 {
+			itr.copyStats()
+		}
 
 		return &itr.point
 	}
 }
 
+// copyStats copies from the itr stats buffer to the stats under lock.
+func (itr *stringIterator) copyStats() {
+}
+
 // Stats returns stats on the points processed.
-func (itr *stringIterator) Stats() influxql.IteratorStats { return itr.stats }
+func (itr *stringIterator) Stats() influxql.IteratorStats {
+	itr.statsLock.Lock()
+	stats := itr.stats
+	itr.statsLock.Unlock()
+	return stats
+}
 
 // Close closes the iterator.
 func (itr *stringIterator) Close() error { return nil }
@@ -1138,7 +1204,9 @@ type booleanIterator struct {
 	m     map[string]interface{} // map used for condition evaluation
 	point influxql.BooleanPoint  // reusable buffer
 
-	stats influxql.IteratorStats
+	statsLock sync.Mutex
+	stats     influxql.IteratorStats
+	statsBuf  influxql.IteratorStats
 }
 
 func newBooleanIterator(name string, tags influxql.Tags, opt influxql.IteratorOptions, cur booleanCursor, aux []cursorAt, conds []*bufCursor, condNames []string) *booleanIterator {
@@ -1150,10 +1218,11 @@ func newBooleanIterator(name string, tags influxql.Tags, opt influxql.IteratorOp
 			Name: name,
 			Tags: tags,
 		},
-		stats: influxql.IteratorStats{
+		statsBuf: influxql.IteratorStats{
 			SeriesN: 1,
 		},
 	}
+	itr.stats = itr.statsBuf
 
 	if len(aux) > 0 {
 		itr.point.Aux = make([]interface{}, len(aux))
@@ -1189,10 +1258,13 @@ func (itr *booleanIterator) Next() *influxql.BooleanPoint {
 
 		// Exit if we have no more points or we are outside our time range.
 		if itr.point.Time == tsdb.EOF {
+			itr.copyStats()
 			return nil
 		} else if itr.opt.Ascending && itr.point.Time > itr.opt.EndTime {
+			itr.copyStats()
 			return nil
 		} else if !itr.opt.Ascending && itr.point.Time < itr.opt.StartTime {
+			itr.copyStats()
 			return nil
 		}
 
@@ -1212,14 +1284,28 @@ func (itr *booleanIterator) Next() *influxql.BooleanPoint {
 		}
 
 		// Track points returned.
-		itr.stats.PointN++
+		itr.statsBuf.PointN++
+
+		// Copy buffer to stats periodically.
+		if itr.statsBuf.PointN%statsBufferCopyIntervalN == 0 {
+			itr.copyStats()
+		}
 
 		return &itr.point
 	}
 }
 
+// copyStats copies from the itr stats buffer to the stats under lock.
+func (itr *booleanIterator) copyStats() {
+}
+
 // Stats returns stats on the points processed.
-func (itr *booleanIterator) Stats() influxql.IteratorStats { return itr.stats }
+func (itr *booleanIterator) Stats() influxql.IteratorStats {
+	itr.statsLock.Lock()
+	stats := itr.stats
+	itr.statsLock.Unlock()
+	return stats
+}
 
 // Close closes the iterator.
 func (itr *booleanIterator) Close() error { return nil }